Move code from do_submit.php to ProcessSubmission.php
